+++ viewer_widget/viewer/panels/clogfilterpanel/clogfilterpanel.h
@@ -18,9 +18,9 @@ public:
     explicit ClogFilterPanel(QWidget *parent = nullptr);
     ~ClogFilterPanel();
 
-    void setClusterRange(const std::vector<double> &vector);
+    void setClusterRange(const std::vector<size_t> &vector);
     void setTotRange(const std::vector<double> &vector);
-    void setTotRangeFull(const std::vector<double> &vector);
+    void setTotRangeFull(const std::vector<float> &vector);
 
     double getClusterBegin() const;
     double getClusterEnd() const;
